Split BackgroundTaskQueueTests.swift into focused test files
569-line file exceeded SwiftLint's 500-line limit. Split into:
- MockGitExecutor.swift: shared MockSimpleGitExecutor used across suites
- BackgroundJobStateTests.swift: BackgroundJobState enum unit tests
- BackgroundTaskQueueTests.swift: queue core/stress/integration tests
- WorktreeListViewModelSelectionTests.swift: bulk-remove ViewModel tests

All existing tests preserved; WorktreeManagerTests retains its own
local mock with a different stubbing API.
